Solve problem set 3
Added solutions for the rest of the problems for problem set 3, along
with the extra credits. Also refactored code slightly so that the
executable entry point in main.cpp is less cluttered. Also added support
for seeding the random number generator in problem 1b with a hexadecimal
string for consistent matrix generation between runs - for a given seed,
the same constraint points and test points will be selected so that the
generated calibration parameters are the same every time.