Problem Set 3 (#6)
* Add framework for problem set 3

Added functions for parsing points from the text files provided in
problem set 3 into vectors of OpenCV matrices. Also set up the
configuration for ps3 and updated the Dockerfile to install Boost.

* Track JPEG files in LFS

* Add Eigen3 to Docker image

* Install Eigen correctly this time

* Compute projection matrix

Added solutions for computing projection matrix with normal least
squares and singular value decomposition.

* Solve problem 1

* Solve problem set 3

Added solutions for the rest of the problems for problem set 3, along
with the extra credits. Also refactored code slightly so that the
executable entry point in main.cpp is less cluttered. Also added support
for seeding the random number generator in problem 1b with a hexadecimal
string for consistent matrix generation between runs - for a given seed,
the same constraint points and test points will be selected so that the
generated calibration parameters are the same every time.

* Add writeup and output files